user-side-service/requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
commonmark==0.9.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
pony==0.7.16
priority==2.0.0
psycopg2-binary==2.9.5
Pygments==2.14.0
python-dotenv==0.21.0
requests==2.28.2
rich==13.1.0
six==1.16.0
SQLAlchemy==1.4.46
toml==0.10.2
tomli==2.0.1
typing-extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wsproto==1.2.0
zipp==3.11.0
